Added scrolling to the text
Can now scroll the text with the up and down arrow keys. Need to work on getting the text to clamp properly, as it is a bit off currently. Also the text now toggles on and off. This feature is temporary, I just wanted to figure out if having two different "states" was possible. I want to work on getting this code view state to appear when the player is trying to unlock a door