Now parsing xml to create and read code blocks
The blocks of code now come from an external xml file. This works quite well for single words or lines of code.
For mutliline segments, the box used for detection can end up a little off because of how the box is calculated. This can be remedied by added whitespace to the xml file. This issue and roundabout solution can be seen in test2.xml