Added some transitioning between levels, fixed up a movement bug, worked on door placement, worked on some culling
I've added a transition animation between diferent parts of the maze.
I have a way to create doors to each part of the maze, though it it pretty basic right now.
Found a bug in the movement code and I believe it is fixed. Also a few more optimizations.
Started working on having different blocks of code show up for the different levels, though they are all currently placeholders
Added a simple and not very good culling of the level objects that are not in the current maze. This can definitely be improved. Note that if it is decided to still draw objects in mazes that the player is not in, the colliders of anything in another maze absolutely should be turned off